BaseSqlTableModel: disable undesirable sorting
This commit fixes undesirable tracks sorting in tables.
The mentioned situation could have occurred, if the stars rating
was changed for some tracks and then after the action
of removing tracks, adding tracks or filtering, the tracks
were sorted.

So, when removing tracks now tracks are kept at the same position
for a user. With adding tracks the new tracks are added at the end
of the tracks table. The automatic sorting is kept when it is changed
between track tables, for example between crates and tracks.

Fixes: #1387373 ("Remove a track from a crate triggers sorting")
Reported-by: Jean Claveau