Library: emit showTrackmodel when activating feature root with tracks table
This makes the track menu 'Select track in library' action work in feature root views that
have a tracks table (AutoDJ, Analyze), though it's not working in Browse and Recording
because the BrowseModel does currently not support getTrackRows().

Issue still unresolved:
When triggering this section while a feature root view with no tracks table is active
the track will be selected in the previously selected tracks table.