Replace using to global cursor stack by using the widget cursor.
This should be a workaround for the missing cursor, that happens on MacOS probably due to a missing release event in overload situations. lp196978
If this happens again, only one knob is affected instead of whole Mixxx and it heals itself by another mouse click because it is not longer stacking up.