PitchShiftEffect: basic structure with one knob
This commit adds a new independent pitch shift effect.
The pitch shift effect contains one knob for setting up
the pitch of a sound. The knob has a fixed range for one octave
up or down. The knob works in linear continuous mode.

The known issue is, that RubberBand library processing latency
needs to be taken into account for different pitch values.
Based on that the effect works with an amount of latency.