PitchShiftEffect: fix destructor error
This commit fixes destructor exception specification error
and refactors code. The code changes are based on the discussion
for the previous PitchShiftEffect commit. Major changes in this commit
are: fix destructor explicit exception specification error, complex
methods movement from the header file into the .cpp file
for PitchShiftGroupState class, use lambda expression for the pitch
value assignment, remove unnecessary casts, add the QStringLiteral
to the getId method and the PitchShiftEffect the class was set
as final.