feat(gh-actions): exempt "needs review" PRs from labelled stale
IMO PRs labelled "stale" mean that the contributor has abandoned the PR.
For many PRs that are not abandoned but simply not being reviewed because
of our reviewer shortage, this is not true. In this case, the PR should be
labelled "needs review" and the "stale" label should not be applied.