modplug: restore settings on load
The settings restoration is currently done in the preferences
dialog. This was dropped accidentally with the introduction of
coreservices in #3446.
Strictly speaking it would be cleaner to not rely on a temporary
widget here. For now this enures modplug tracker files can be
loaded and play correctly after application startup.