EngineEffectsDelay: frames instead of samples
The commit changes the type of a value which represents
propagated delay. This commit changes the code to use frames
for propagating delay instead of samples. The reason, why propagating
frames is better than samples is that it doesn't depend on the number
of channels. However, the inner EngineEffectsDelay structure uses
samples for calculations still, so the number of samples is recalculated
by propagated number of frames and a current number of channels.
To clean up the workflow, the input buffer isn't used as the output
buffer too and the temporary buffer replaces it. The size of the delay
buffer is known in the compile time.