EngineEffectsDelayTest: add delay handling tests
This commit adds tests for the EngineEffectsDelay.
The EngineEffectsDelayTest contains nine tests including two tests
for expected failure, which means debug assertion. Then, it contains
tests for delaying buffers, crossfading and copying for zero delay
value. The EngineEffectsDelayTest assumes, that the Mixxx works
with a stereo signal. In case of future changes, tests must be updated.