EngineEffectsDelayTest: add stereo assertion
This commit adds the static_assert for the EngineEffectsDelayTest.
The signal in the input buffer has to be a stereo signal for the tests
to be successful. It's due to the EngineEffectsDelay working
with the channel count information and all tests are written
for the stereo signal, with whom the Mixxx works.