EngineEffectsDelay: inheritance from EngineObject
This commit changes the class from which the EngineEffectsDelay
inherits. The EngineObjectConstIn is replaced by EngineObject.
The fundamental difference is that the EngineObject works with only
one buffer for input and output so far as EngineObjectConstIn works
with two buffers. One for input and one for output. The one buffer
version which works in place fits better for the usage
in the EngineEffectChain and is better from a performance perspective.