EngineEffectsDelay: add benchmarks for testing
This commit adds benchmarks for testing EngineEffectsDelay.
The benchmarks test different cases. First for the zero delay,
then for delay smaller than the buffer size, for delay greater
than the buffer size, for crossfading between two different delays
and compared to the same situation without crossfading
and with the same delay.