EngineEffectsDelay: modulo calculation doc
The commit adds documentation for the modulo calculation.
The mentioned documentation explains the usage of the addition
of the size of the delay buffer in the calculation, which is used
for the modulo operator as well (as a divisor for division).
From a mathematical point of view, this addition can be removed.
Anyway, from the cpp point of view, the delay buffer size addition
is necessary based on an otherwise negative result using the cpp
modulo operator.