EngineDelay: modulo calculation doc
This commit adds documentation for the modulo calculation
in the EngineDelay::process. The mentioned documentation explains
the usage of the addition. The added value is the size
of the delay buffer, which is used for the modulo operator as well
(as a divisor for division). From a mathematical point of view,
this addition can be removed. Anyway, from the cpp point of view,
the delay buffer size addition is necessary based on an otherwise
negative result using the cpp modulo operator.