RingDelayBuffer: add a new ring buffer
This commit adds a new ring buffer. This ring buffer serves
for handling a delay. For the first use case was specially created
for the EngineEffectsDelay, which handles the delay, of the effects
(in a chain). The ring delay buffer allows moving
with the reading position subject to certain rules. Another difference
between the classic ring buffer is, that the ring delay buffer offers
to read zero values which were not written by using the write method
and write position. Both of these two specific properties are based
on the cross-fading between changes of two delays
and the classic ring buffer cannot be used.