RingDelayBufferTest: add tests for a ring buffer
The commit adds tests for the ring buffer for handling delay.
The test set includes tests for testing the RingDelayBuffer::isEmpty,
RingDelayBuffer::isFull, RingDelayBuffer::clear, then tests
for checking the number of available items for reading and writing
and at last tests for reading and writing without skipping the position
with the read position and including skip with the read position
on both sides and with both variants (circle around / not circle
around).