RingDelayBuffer: add original license
The commit adds the original license for the ring buffer
from the Portable Audio I/O Library. The ring buffer
from the pa_ringbuffer.c was used as a template for some parts of code
in RingDelayBuffer. Based on that, the original license is added
to cpp file and header. In the header, the modified and added functions
are described.