EngineEffectsDelay: add kMaxDelayFrames constant
This commit adds a kMaxDelayFrames constant for clamping the upper bound
of delay. This constant represents the maximum possible delay,
which can be provided. However, the size of the delay buffer size
is greater than about one frame (multiply by the number of channels
to convert to samples). This difference by one frame ensures,
that it can be read only the oldest sample value and
without the possibility of a situation where the channels are switched.
The maximum delay cannot be the same as the delay buffer size.
If this situation were to be allowed, the first read sample
will be the newest and the next samples will be from the oldest.