EngineFilterDelay: add assertions
This commit adds two assertions. The first one is a static_assert
which tests if the templated SIZE is divisible
by the number of channels. The second one is VERIFY_OR_DEBUG_ASSERT
in EngineFilterDelay::setDelay. It tests if the number of delay samples
is divisible by the number of channels. If the delay size
is not aligned, the unaligned samples are cropped (rounded
to the previous multiple of the number of channel count).