PitchShiftEffect: report initial latency
This commit adds the reporting of the Pitch shift effect latency.
The reported value is initial latency based on prefilling the RubberBand
to avoid dropouts based on RubberBand processing. The known issue is,
that the reported value is valid for the initial pitch settings only.
After dynamic pitch changes, the value has to be changed
and the current implementation does not work properly
for these use cases for now.