RingDelayBuffer: skip for invalid number of items
This commit changes the situation if the number of samples to read
is greater than the read available samples, or when
the number of samples to write is greater than the write available
samples space. In both situations, the code does not process the command
and just skips the invalid request.