RingDelayBuffer: read with provided delay
This commit mainly optimizes the RingDelayBuffer::read.
The read position is not newly stored and the number of delay items
is provided directly into RingDelayBuffer::read. The read position
is calculated using the write position, the number of items to read
and mentioned delay. Based on that the unused methods are deleted,
which are RingDelayBuffer::moveReadPositionBy, RingDelayBuffer::isFull,
RingDelayBuffer::isEmpty, RingDelayBuffer::getReadAvailable
and RingDelayBuffer::getWriteAvailable. Due to all code changes
some tests and benchmarks are removed and some are upgraded.