PitchShiftEffect: add wide range option
This commit adds the Wide Range option for the Pitch shift effect.
This option doubles the scale of the pitch knob, which means,
that with the toggle on, the knob range is not in -1 octave
and +1 octave, but in doubled range, so in range -2 octaves
and +2 octaves. Based on the changes, the knob now works
for both situations in the exponential scale of the power of 2.