PitchShiftEffect: add range mode option
This commit adds the Range option for the Pitch shift effect.
This option changes the range of the pitch knob. There are three
possible options based on the knob position. The neutral mode (knob
in the default position) keeps the range from the lowest pitch
to the highest pitch. The positive mode (knob in the maximum position)
keeps the range from the default pitch to the highest pitch.
The negative mode (knob in the minimum position) keeps the range
from the default pitch to the lowest pitch. Every mode works correctly
with set the Wide Range option too.