Dispose Track::refreshCoverImageDigest() reading covers from the main thread.
This fixes https://bugs.launchpad.net/mixxx/+bug/1905124

After a cover is read by the CoverArtDelegate via a worker thread, it tries to update the Cover Image Digest via a Track object. During creation of this Track object, the cover is read again this time from the main thread, creating the scrolling issue.