PitchShiftEffect: add a formant preserving option
This commit adds the Formant Preserving option for the Pitch shift
effect. This option changes the settings of formant processing
using the RubberBand library.  The default option doesn't use
any special formant processing. When the option is turned on,
the handling of the formant shape is activated and the spectral envelope
of the unshifted signal is preserved.