math: clean up sgn function
This commit cleans the sgn function which implements the sign function
as widely known. The static assert for checking, if a type
is an arithmetic, is replaced using the required keyword
for the same situation. Newly introduced is checking if the type
is unsigned. If yes, the calculation can be simplified. Implicit casts
are replaced with explicit casts.