RingDelayBuffer: handle invalid copy alignment
The commit adds handling structure for the situation when
the number of items to copy causes, that both buffers have to cross
their upper bounds and circle around at least once. For the current
RingDelayBuffer::copyRing usage this situation is not required, so,
the VERIFY_OR_DEBUG_ASSERT is added.