PitchShiftEffect: add id literals into a namespace
This commits creates literal constants in the anonymous namespace.
The constants values are the same as the manifest
parameter identifiers. The constants are created to avoid grammar bugs
due to the value has to be used twice on the other code places. Due to
the literal being used, the segmentation fault has occurred
without previous warnings or errors if the entered identifier names
for the specific parameter are not the same.