RingDelayBuffer: avoid uninitialized memory
This commits tries to avoid the uninitialized memory due to creating
the variable before the if-else statement without assigning
the default value. One solution could be assigning the default value,
anyway, some IDEs can warn that the variable is unused and primarily
it will hide real errors like that the variable is not set
in one branch. Based on that, the lambda solution is used.