CircularBuffer: number of items same as the size
This commit improves the implementation of the circular (ring) buffer.
For the previous implementation due to inner calculations
of if the buffer was full or empty it was not possible
to completely fill the buffer with the same amount of samples as was
the size provided during the construction. For the previous version,
it was possible to store only "size - 1" samples. The new version allows
for storage exactly the same number of items as was the size provided.