PitchShiftEffect: introduce the pull model
This commit introduces the RubberBand pull model implementation
in the Pitch shift effect. The pull model is inspired
by the implementation from
engine/bufferscalers/enginebufferscalerubberband.cpp. The pull model
can be implemented for the effect due to the input ring buffer
being implemented. The input ring buffer pre-cached the input buffer
for the RubberBand. If the size of the pre-caching ring buffer
is enough, the situation that the RubberBand will require
more input sample frames that were stored in the mentioned ring buffer,
could not occur.