EngineEffectsDelay: introduce ring delay buffer
This commit introduces the RingDelayBuffer and replaces the previous
processing approach. Based on that, the implemented crossfade
is replaced with a function from SampleUtil, concretely
SampleUtil::linearCrossfadeBuffersOut. Dependent tests are upgraded
because RingDelayBuffer uses the fading-in, for the first written chunk
of data, and the crossfading is used in stereo mode.