Add beatlooproll effect to DDJ 400
I love the beat loop roll effect.  This change adds a way to use this
with the Pioneer DDJ-400 controller.

The `PAD FX1` mode for the pads was previously not assigned to
anything.  With this change, it is assigned to the beat loop roll
effect.  This is similar to actually how some of the pads work in
Rekordbox when in this mode, afaik.

I start at the lower beat loop roll length, since longer roll sizes
are actually not that useful (one needs to hold the button for the
roll).