Fix condition in JACK backend check
`m_deviceInfo->hostApi` is an index in the list of supported APIs on
this platform. So on Linux 0 is ALSA, 2 is JACK. This is also mentioned
in the comment on the `hostApi` field. We should be checking the device
type ID instead, as is done elsewhere in the code.

This fixes the buffer size when using JACK clients. With JACK the buffer
size should be coming from the server (and PortAudio chops it up into
1024 sample blocks if it's larger). Before this change it would use the
now grayed out latency value set for the previous audio backend before
switching to JACK (defaulting to 1024 samples).