Replace library home/end key handlers with cursors
By simply overriding the cursor movement function instead of adding an
explicit key event the home and end keys in the library table view now
interact as expected with other hotkeys and cursor actions. Shift+End
now select all items through the end of the list, and Ctrl+Home jumps to
the first item without selecting it so it can be toggled on and off with
Space.