Traktor Kontrol S2MK3: Use FX select buttons to set quick effect presets
This repurposes the "FX SELECT" buttons on the controller to change the
loaded quick effect preset rather than toggle effect chains. This allows
them to be used in conjunction with the filter knobs on either side.

Pressing a button will load the preset at that index of the quick effect
preset list (pressing button 1 will load the first preset, 2 the second,
etc.). Holding down shift while pressing a button will only change the
preset for the deck whose shift button was used. Buttons will light up
accordingly, and if two different presets are active on one deck each,
both of the buttons will light up.

One issue now is that changing the effect in the GUI won't change the
LED status of the buttons, but I'm not sure how feasible this would even
be to implement.