Replace m_framesPerBuffer uses with runtime size
In any operation that works with audio buffers passed from PortAudio.
This is needed because PortAudio may pass us buffers that are smaller or
larger than the size we request. And when using JACK the requested size
is 0, which will cause PortAudio to pick a buffer size matching the JACK
server.