Replace manual memory management in rubberband
These functions still use raw pointers so there's a bunch of stuff
that's difficult to catch that can still go wrong, but this at least
removes the need for having to manually deallocate these structures. And
this class didn't adhere to the rule of five, so it was very easy to
misuse it and cause double frees and use after frees.