fix(serato): Fix resetting track colors on metadata reimport
When metadata for a file without Serato tags is reimported, the track
color is reset. This happens, because there is currently no way to
distinguish the case where no track color is set because the file does
not contain any Serato tags from the case where the file contains Serato
tags and the track color has been set to "no color" in Serato (i.e. the
stored color value is `0xFFFFFF`.

This fixes the issue by replacing the `RgbColor::optional_t` value with
a `std::optional<RgbColor::optional_t>` to separate the 3 cases:

1. If the outer optional is `nullopt` , there is no track color present in
  the tags.
2. If the inner optional is `nullopt`, there is a track color present in
  the tags, and that color is "no color".
3. If the inner option holds a value, there is a track color with that
   RGB color value present in the tags.

Fixes #11213.