fullscreen: ignore no-op QWindowStateChangeEvent to not change menu bar
No-op QWindowStateChangeEvent can occur if another window (preferences, About, ...) is shown above
MixxxMainWindow. The event's oldState() is QWindowNoState even though the window is and was fullscreen.
Ignore these event to not recreate and reconnect the menu bar unnecessarily.
Also, that avoids flickering, at least on desktops that feature a global menu (when going fullscreen
the menu is moved to the main window and hidden immediately = rapid resizing of window content)