Only Import/Export Serato compatible cue indexes.
Export and Import Serato Loop cues with a fixed offset off 8. Skip all cue entries that are not usable in Serato. Do not shift cues.
This keeps the cues identifiable by there index resp. the position in the grid. This may match the position in a track or a position schema of the user.