fix OpenGL version detection
This allows OpenGL waveforms to be used with the Wayland Qt
Platform Abstraction on KDE and Weston. Though GL_MAJOR_VERSION and
GL_MINOR_VERSION reported much higher versions (4.6 when tested on
Intel UHD Graphics 770 (ADL-S GT1)), QSurfaceFormat reports 2.0 on
KDE Wayland and Weston (somehow it reports 2.1 on GNOME Wayland
though) and 2.1 with the Xcb QPA. QSurfaceFormat does not report
higher versions if they are requested in the OpenGLWindow
constructor.