MixxxApplication: Use XCB plugin for static Qt builds on Linux
Currently, Mixxx cannot be built at all using static Qt and will always
error at compile time with a message indicating the a plugin for the
current platform is missing.

While using the system-packaged Qt is generally the preferred way, there
are use cases for statically linking Qt (e.g. portable or embedded
builds), therefore it would be nice if we could at least support
building Mixxx this way.

Unfortunately, Linux does not have a single display API, a list can be
found in the Qt documentation:

    https://doc.qt.io/qt-6/qpa.html

This patch chooses to hardcode the X11/XCB plugin for static Qt builds,
which is certainly not ideal, especially since Wayland seems to
increasingly be the preferred backend, but seems to be the only plugin
supported by vcpkg's `qtbase`:

    https://github.com/mixxxdj/vcpkg/blob/2.5-rel/ports/qtbase/portfile.cmake