fix(controller): mass replace outdated usage idioms of `engine.beginTimer`
Due to the switch to QJSEngine in Mixxx 2.4 the previous API
where the `this` of the callback function was set to the `this` in the
calling context of the `beginTimer` is not possible to retain
(nor does it make much sense anymore). For this reason this commit
fixes all controllers where this behavior was assumed and updates
the rest of them to use arrow functions instead of anonymous function
expressions or strings.