Sync Lock: protect against syncing phase to a stopped leader
Fixes #11788

Propagate playing/audible state to the bpmcontrol.
If a Leader is not playing, we should not use it for sync adjustment or phase seeking, because that causes audible artifacts.

Moreover, we surrender an explicit leader if it is stopped.
In the past we had thought of explicit leader as a persistent user choice, but in reality it is more likely to be a temporary setting which the user would like us to fix when it stops being relevant.
There are lots of issues with having a stopped deck as leader, so this seems like the right choice.