fix(controllers): ensure callbacks have appropriate `this`
This should fix most mappings that were previously broken,
because they assumed when a function was passed as a callback
to `engine.makeConnection` or `engine.beginTimer` that the
`this` argument referred to the same `this` as when the
function was defined (eg when defined as a "method" on a class).
In reality, any function executed via [`QJSValue::call`](https://doc.qt.io/qt-5/qjsvalue.html#call)
(which is the case with callbacks executed via
`ControllerScriptEngineBase::executeFunction`), will have its
`this` argument set to the globalObject!

This commit fixes some of these issues in known cases (https://github.com/mixxxdj/mixxx/issues/12460),
potential places introduced as part of (https://github.com/mixxxdj/mixxx/pull/12401)
as well as some other potential spots that were easy to grep for.