Implement cpu pinning and cgroup cpuset support
Add command line argument to specify a cpuid and cpuset cgroup name for the engine.
The engine tries at startup to move itself to the cpuset provided or set the cpu affinity.
This will allow a wrapper program to free a core for the engine, minimizing latency jitter or CPU starvating.