Waveform/Spinny scratching: fix loop wrap-around issue
PositionScratchController::process() now knows if the play position was wrapped around (and
how often) since the last processing. This avoids an infinite wrap-around 'loop', even after
the mouse didn't move.

Previously, PositionScratchController didn't know if the play position was wrapped around, so it
interpreted the raw (huge) sample delta as play pos move and calculated a huge rate, which was
then applied by the engine, resulting in a (now true) long sample distance in the next run,
which caused another loop wrap-around, and so on...