StarDelegate (1/4): Detect all cases where the mouse leaves the editor control
...and close the persistent editor in those cases.

This is important for keeping the our internal bookkeeping
in sync with the Qt framework state - the former is required
to work around some quirks related to the two ways of
interacting with the StarEditor (by mouse and by keyboard)
and how this interacts with the assumptions of QTableView.

                ====================
                = More information =
                ====================

The StarRating control can be edited via mouse, keyboard, or
both at the same time. The following transitions are currently
handled correctly (marked by ✓) resp. affected by this change
(marked by =>):

       [✓] "Not editing"       -> "Edit via mouse"          (on mouseEnter)
       [✓] "Not editing"       -> "Edit via keyboard"       (on editKeyPressed)

[ ] => [✓] "Edit via mouse"    -> "Not editing"             (on mouseLeave)
       [ ] "Edit via mouse"    -> "Edit via keyboard+mouse" (on editKeyPressed)

       [✓] "Edit via keyboard" -> "Not editing"             (on exitKeyPressed)
       [✓] "Edit via keyboard" -> "Edit via keyboard+mouse" (on mouseEnter)

       [ ] "Edit via keyboard+mouse" -> "Edit via mouse"    (on exitKeyPressed)
[ ] => [~] "Edit via keyboard+mouse" -> "Edit via keyboard" (on mouseLeave)