AutoDJProcessor: Add control objects for Auto DJ queue length and duration
The queue_duration and queue_tracks controls in the [AutoDJ] group
can be used to display the information about the Auto DJ queue
in custom skins.

The current calculation method for the queue_duration control
only considers the total duration of the tracks by themselves,
and does not take the transition times between tracks into account.