AutoDJProcessor: Feature: Add control objects for Auto DJ queue status
The time_remaining and tracks_remaining controls in the [AutoDJ]
can be used to display the information about the Auto DJ queue
in custom skins.

The current calculation method for the time_remaining control
only considers the total duration of the tracks by themselves,
and does not take the transition times between tracks into account.