refactor(util): Modernize & shrink ScopedTimer internals
Instead of emulating an optional by use of a pointer and
placement-`new` construction, use a `std::optional` directly.
Also remove `m_cancel` as the empty optional is sufficient for
indicating a cancelled timer. Removing these members shrunk the
memory footprint from 72 to 56 bytes, reducing the memory overhead
of ScopedTimer compared to the underyling Timer to 8 bytes
(Timer currently being 48 bytes large).
While these improvements are certainly nice, the purpose of this
PR is the improved underlying code. However, there is still
lots that could be improved in the future.