feat: make `parented_ptr` `std::move`-able
Since parented_ptr is "just" a marker type for an owning pointer
(similar to `std::unique_ptr`) it makes sense to enable it to
be moved. This makes it possible to use parented_ptr in places
where a parented QObject is constructed, but not yet transfered
to its owner (this is the case in factory functions).