Library: Fix: setLibraryFocus should work even if the window is not currently focused
This scenario occurs e.g. when calling setLibraryFocus from a menu action.
The popup window has already been closed when QAction::triggered occurs,
but the main window has not yet gotten its focus back.

As per the documentation of QWidget::setFocus(), when the containing
window is not active, the focus change request is delayed until the
window becomes active - which is exactly what we want in our case.