Fix double reference detection.
The ID is the unique ID which is always returned. When finding a track by canonical location, without an ID, verify that the location stored in the database which can contain symlinks is also matching. If not a different track referencing that same file is already cached and we need to return null to not introduce a data race.

This avoids the issue that wrong track can be returned when using drag and drop.