StarDelegate (2/4): Transition from "mouse edit mode" to "keyboard+mouse edit mode" on edit key pressed
Correctly handle the case of a user pressing the edit key while the
mouse is also hovering over the focused cell.

Without this special handling, the pressing the edit key on a cell
would simply be ignored if the mouse is hovering over that cell,
which would most certainly break user expectations.

                ====================
                = More information =
                ====================

The StarRating control can be edited via mouse, keyboard, or
both at the same time. The following transitions are currently
handled correctly resp. affected by this change:

       [✓] "Not editing"       -> "Edit via mouse"          (on mouseEnter)
       [✓] "Not editing"       -> "Edit via keyboard"       (on editKeyPressed)

       [✓] "Edit via mouse"    -> "Not editing"             (on mouseLeave)
[ ] => [✓] "Edit via mouse"    -> "Edit via keyboard+mouse" (on editKeyPressed)

       [✓] "Edit via keyboard" -> "Not editing"             (on exitKeyPressed)
       [✓] "Edit via keyboard" -> "Edit via keyboard+mouse" (on mouseEnter)

       [ ] "Edit via keyboard+mouse" -> "Edit via mouse"    (on exitKeyPressed)
       [~] "Edit via keyboard+mouse" -> "Edit via keyboard" (on mouseLeave)