feat: always print eslint hook verbose to avoid swallowing warnings
Otherwise the eslint stdout is only printed if eslint actually fails.
In the case of warnings, we don't want pre-commit to fail, but we
still want to bring attention to the users code that something is
suboptimal. In order for that to work the output still needs to be
shown even when the hook does not fail. Setting a hook to be
`verbose` is the only way to make that happen as far as I can tell.