Numark Mixtrack 3, library navigation: add option for Classic and Focus mode
In Focus mode the Browse encoder affects the library widget which currently has keyboard focus.
This requires the Mixxx window to have focus and allows to control the searchbar, too.

In Classic mode the Browse encoder can select items in the sidebar and tracks table only, but independent
of keyboard focus.

Co-authored-by: Swiftb0y <12380386+Swiftb0y@users.noreply.github.com>