Add case-insensitive parsing for Lancelot key notation
When a user types in "key:5a" into the search field, the intent is probably
to search for "key:5A" instead of doing a text-only search. The only case
where this is ambigous is for the d/m (OpenKey) and D/M (Lancelot) suffixes,
where lowercase will imply OpenKey, and uppercase will imply Lancelot.