Fix pitch bend buttons on Hercules DJ Console Mk1
Previously the pitch bend buttons were mapped the same way as the jog
wheels, however this is identical to the jog wheels right below them and
felt a bit useless. This patch maps the buttons back to bumping the
pitch up or down, and adds a setting that allows reverting to the
previous behavior for backwards compatibility.

Signed-off-by: Sam Whited <sam@samwhited.com>