Use ComponentJS for the Twister bindings
This does not change any functionality by default except that the LED
below each button automatically becomes an indicator letting you know
the controller is centered.

It does, however, change the entire binding over to use ComponentJS
which in turn allows us to use two of the side buttons to toggle the
left and right sides of the controllers between decks.

Signed-off-by: Sam Whited <sam@samwhited.com>