(fix) BrowseFeature: don't automatically rebuild tree on each expand
This can cause a crash when jumping to a SidebarBookmark in a currently collpased tree
(or, with many adjustments in sidebar & tree models, at least not give the desired result).

Only build the tree when there are currently no items, e.g. on initial expandation,
or rebuild when `forceRebuild` argument is true.
Else, if there is a mismatch between current and new items, show a refresh icon on the parent.

It's always possible to enforce a tree rebuild with the menu action "Refresh directory tree".