sync: prefer playing inaudible decks over stopped non-sync decks
When selecting a new leader or a sync target, add one more low-priority category of target: playing non-audible decks.
This is useful when the DJ is auditioning tracks in the headphones with nothing going out to the main outputs.
Non-audible decks will only be preferred targets over completely stopped non-sync decks.