Remove useless udev rule
The udev rule in question is a remnant of the old udev rules file
("mixxx.usb.rules") which used GROUP="users" instead of TAG+="uaccess"
(changed in commit c9bb987).  It is rendered useless in the current uaccess
scheme, but the commit author erroneously changed GROUP="users" to
GROUP="uaccess" instead of removing the line.