fix: gcc `-Warray-bounds=` in fidlib by using a flexible member
I carefully reviewed the fidlib code to ensure any `FidFilter`
related size calculations go through the `FF*` macros which
have been adapted to the new `sizeof(FidFilter)` value (which
now excludes the flexible `val` member in the calculation, which
also incidentally simplifies the actual calculation).
Flexible members are a C99 feature and since the fidlib code
essentially hasn't been updated since ever since it has been added
in June 2007, I assume that it wasn't written like this in the
first place so its pre-C99 compatible. This concern is void in
the year 2025.